+++ src/app/(layout-with-navigation)/(simulation)/simulateur/[root]/_components/Simulateur.tsx
@@ -3,9 +3,9 @@
 import Total from '@/components/total/Total'
 import { useDebug } from '@/hooks/useDebug'
 import { useState } from 'react'
-import Form from './Form'
-import Summary from './Summary'
 import Charts from './charts/Charts'
+import Form from './form/Form'
+import Summary from './summary/Summary'
 
 export default function Simulateur() {
   const [isQuestionListOpen, setIsQuestionListOpen] = useState(false)
